% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBI_measurecalculation_normalisation.R
\name{AMBI_measurecalculation_normalisation}
\alias{AMBI_measurecalculation_normalisation}
\title{AMBI_measurecalculation_normalisation
Internal calculation used in calculate AMBI. Makes the sum of the items of a personality trait (reverse item are reversed before) and then normalize them between O and 1.}
\usage{
AMBI_measurecalculation_normalisation(df, v_pos, v_rev)
}
\arguments{
\item{df}{data frame with our data cleaned.}

\item{v_pos}{vector of the items positively correlated with the personality trait we aim to calculate.}

\item{v_res}{vector of the items negatively correlated with the personality trait we aim to calculate.}
}
\value{
a vector of the result of the AMBI we wanted to calculate
}
\description{
AMBI_measurecalculation_normalisation
Internal calculation used in calculate AMBI. Makes the sum of the items of a personality trait (reverse item are reversed before) and then normalize them between O and 1.
}
