% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AMBI.R
\name{calculate_AMBI}
\alias{calculate_AMBI}
\title{calculate_AMBI This function aim to calculate measure for 200 personality scales with 200 items}
\usage{
calculate_AMBI(
  df,
  var_name = "element_name",
  recodage = function(v) {     AMBI::recode_AMBI(v) },
  calculation_function = "normalization_0-1"
)
}
\arguments{
\item{df}{your data containing the answer of participants to 200 items, and gives in result a dataframe with measure of 200 personality scales. The data frame put in entry has either to have columns' names named like AMBI::df_NDA\link{,'ElementName'} or AMBI::df_NDA\link{,'Aliases'} or have the columns in the same order as AMBI::df_example_elementname. A dataframe example as been put under the name of AMBI::df_example_elementname and of AMBI::df_example_aliases, if you use the aliases as variable names.}

\item{var_name}{a character string that should be in c('element_name','aliases') or NULL (by default 'element_name'); element_name: indicate that the variable name of your item are in AMBI::df_NDA\link{,'ElementName'}; aliases: indicate that the variable name of your item are in AMBI::df_NDA\link{,'Aliases'}; NULL: indicate that the variable indexes of your item  are in the same order than the row in AMBI::df_NDA}

\item{recodage}{a function on how to recode your variable if necessary (by default AMBI::recode_AMBI()).}

\item{calculation_function}{a character string in c('normalisation_0-1', 'scale') or a function on how the calculation of the AMBI should be done (by default 'normalisation_0-1'). 'normalisation_0-1' makes the sum of the items of a personality trait (reverse item are reversed before) and then normalize the result between 0 and 1. 'scale' makes the sum of the items of a personality trait (reverse item are reversed before) and then scale them. These functions suppose that your data is coding between 1 and 5 for each items. If you want to add your own function, it must use and have 'df_pos' (dataframe with the column of the items positively correlated with AMBI) et 'df_rev' (dataframe with the column of the items negatively correlated with AMBI) arguments.}
}
\value{
a dataframe combining your dataframe put in entry and the calculated measure of AMBI personality scale
}
\description{
calculate_AMBI This function aim to calculate measure for 200 personality scales with 200 items
}
\references{
T. Yarkoni, "The abbreviation of personality, or how to measure 200 personality scales with 200 items", \emph{Journal of Research in Personality} 44 (2010) 180–198, doi:10.1016/j.jrp.2010.01.002
}
\author{
Léo HENRY
}
