% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculat_big5.R
\name{calculate_big5}
\alias{calculate_big5}
\title{calculate_big5
Calculate the big5 from the personality trait scale calculated with AMBI::calculate_AMBI()}
\usage{
calculate_big5(
  df,
  calculation_method = "normalisation",
  semscoring_method = "normalisation"
)
}
\arguments{
\item{df}{a dataframe with the data issued by AMBI::calculate_AMBI()}

\item{calculation_method}{must be in c('normalisation', 'scaling','semscoring') (default 'normalisation'). semscoring generate a sem using the model of construction of the big five and then use the estimate to calculate the latent variable (the big 5).}

\item{semscoring_method}{if calculation method use the semscoring methods, then it is possible to normaliza the result or scale them. Calculation method/semscoring method must be in c('', scale','normalisation'), (default 'normalisation').}
}
\value{
if calculation_method in c('normalisation', 'scaling'), then a dataframe with the big5. If calculation_method is 'semscoring', then a list with the SEM and a dataframe with the big5.
}
\description{
calculate_big5
Calculate the big5 from the personality trait scale calculated with AMBI::calculate_AMBI()
}
