% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_normalisation.R
\name{f_normalisation}
\alias{f_normalisation}
\title{f_normalisation
normalise a vector between 0 and 1 (it does not take into consideration standard-deviation)}
\usage{
f_normalisation(v)
}
\arguments{
\item{v}{a numerical vector}
}
\value{
a normalized numerical vector
}
\description{
f_normalisation
normalise a vector between 0 and 1 (it does not take into consideration standard-deviation)
}
