% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_AMBI.R
\name{recode_AMBI}
\alias{recode_AMBI}
\title{Recode AMBI survey's question
This function will clean the results to be sure that in case of abcense of answer the value will be replace by NA}
\usage{
recode_AMBI(v_AMBI)
}
\arguments{
\item{v_AMBI}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
Recode AMBI survey's question
This function will clean the results to be sure that in case of abcense of answer the value will be replace by NA
}
