% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semscoring_latent.R
\name{semscoring_latent}
\alias{semscoring_latent}
\title{semscoring_latent
method to score latent variable found in sem. The method use the estimate to determine the importance of each variable on the latent variable.}
\usage{
semscoring_latent(fit, df, calculation = "normalisation")
}
\arguments{
\item{fit}{sem object from the lavaan package}

\item{df}{dataframe on which the the sem was calculated}

\item{calculation}{calculation method, must be in c('normalisation', 'scale')}
}
\value{
a data frame with the score of the latent variable
}
\description{
semscoring_latent
method to score latent variable found in sem. The method use the estimate to determine the importance of each variable on the latent variable.
}
